trie-d -> trie-db-old, trie-db-ext -> trie-db
