--- packages/trie-db-ext/README.md
@@ -1,3 +1,0 @@
-# @polkadot/trie-db-ext
-
-A re-implementation of a Patricia Trie. Unlike other implementations, this allows for the specification of a hash function as well as encoder/decoder and operates in a sync fashion by default.
