--- packages/trie-db-ext/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@polkadot/trie-db-ext",
-  "version": "0.28.25",
-  "description": "This is an implementation of the modified merkle patricia tree as speficed in Ethereum's yellow paper, adpated for Polkadot",
-  "main": "index.js",
-  "keywords": [
-    "Polkadot"
-  ],
-  "author": "Jaco Greeff <jacogr@gmail.com>",
-  "maintainers": [
-    "Jaco Greeff <jacogr@gmail.com>"
-  ],
-  "contributors": [],
-  "license": "ISC",
-  "engines": {
-    "node": ">=6.4"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/polkadot-js/common.git"
-  },
-  "bugs": {
-    "url": "https://github.com/polkadot-js/common/issues"
-  },
-  "homepage": "https://github.com/polkadot-js/common/tree/master/packages/trie-db-ext#readme",
-  "dependencies": {
-    "@babel/runtime": "^7.0.0-rc.1",
-    "@polkadot/trie-hash": "^0.28.25",
-    "@polkadot/util": "^0.28.25",
-    "@polkadot/util-crypto": "^0.28.25",
-    "@polkadot/util-rlp": "^0.28.25"
-  },
-  "devDependencies": {
-    "@polkadot/db": "^0.28.25",
-    "ethereumjs-testing": "1.0.4"
-  }
-}
