+++ packages/trie-db-old/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@polkadot/trie-db-old",
+  "version": "0.28.25",
+  "description": "This is an implementation of the modified merkle patricia tree as speficed in Ethereum's yellow paper, adpated for Polkadot.",
+  "main": "index.js",
+  "maintainers": [
+    "Jaco Greeff <jacogr@gmail.com>"
+  ],
+  "contributors": [
+    "mjbecze <mjbecze@gmail.com>",
+    "Aaron Kumavis <http://aaron.kumavis.me/> (https://github.com/kumavis)"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/polkadot-js/common.git"
+  },
+  "bugs": {
+    "url": "https://github.com/polkadot-js/common/issues"
+  },
+  "keywords": [
+    "merkle",
+    "radix",
+    "trie",
+    "ethereum"
+  ],
+  "license": "MPL-2.0",
+  "dependencies": {
+    "@babel/runtime": "^7.0.0-rc.1",
+    "@polkadot/trie-hash": "^0.28.25",
+    "@polkadot/util": "^0.28.25",
+    "@polkadot/util-crypto": "^0.28.25",
+    "@polkadot/util-rlp": "^0.28.25",
+    "@types/async": "^2.0.49",
+    "@types/semaphore": "^1.1.0",
+    "async": "^2.6.1",
+    "encoding-down": "^5.0.3",
+    "level-ws": "0.1.0",
+    "levelup": "^3.1.1",
+    "memdown": "^3.0.0",
+    "semaphore": ">=1.0.1"
+  },
+  "devDependencies": {
+    "ethereumjs-testing": "1.0.4"
+  }
+}
