+++ packages/trie-db-old/src/BaseTrie.ts
