+++ packages/trie-db-old/src/Node.ts
