+++ packages/trie-db-old/src/encoder/decode.ts
