+++ packages/trie-db-old/src/encoder/encode.ts
