+++ packages/trie-db-old/src/encoder/index.ts
