+++ packages/trie-db-old/src/index.ts
@@ -1,0 +1,23 @@
+// Copyright 2015-2018 authors & contributors of https://github.com/ethereumjs/merkle-patricia-tree
+// Copyright 2017-2018 @polkadot/trie-db authors & contributors
+// This software may be modified and distributed under the terms
+// of the MPL-2.0 license. See the LICENSE file for details.
+
+import { HashFn } from './types';
+
+// import logger from '@polkadot/util/logger';
+import CheckpointTrie from './CheckpointTrie';
+import { prove, verifyProof } from './proof';
+
+// const l = logger('trie');
+
+export default class Trie extends CheckpointTrie {
+  static prove = prove;
+  static verifyProof = verifyProof;
+
+  constructor (db: any, root?: Uint8Array, hashing?: HashFn) {
+    super(db, root, hashing);
+
+    // l.debug('Created Trie');
+  }
+}
