+++ packages/trie-db-old/src/nibbles/index.ts
