+++ packages/trie-db-old/src/nibbles/isEqual.ts
