+++ packages/trie-db-old/src/nibbles/matchingLength.ts
