+++ packages/trie-db-old/src/nodeFactory.ts
