+++ packages/trie-db-old/src/proof/prove.ts
