+++ packages/trie-db-old/src/proof/verifyProof.ts
