+++ packages/trie-db-old/src/streams/ScratchRead.ts
