+++ packages/trie-db-old/src/types.d.ts
