+++ packages/trie-db-old/src/util/isRawNode.ts
