+++ packages/trie-db/src/util/key.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/trie-db-ext authors & contributors
+// Copyright 2017-2018 @polkadot/trie-db authors & contributors
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
